Support for reuse of channels
